+++ src/Glpi/Console/Application.php
@@ -374,14 +374,6 @@ class Application extends BaseApplication
      */
     private function initSession()
     {
-
-        if (!Session::canWriteSessionFiles()) {
-            throw new \Symfony\Component\Console\Exception\RuntimeException(
-                sprintf(__('Cannot write in "%s" directory.'), GLPI_SESSION_DIR)
-            );
-        }
-
-        Session::setPath();
         Session::start();
 
        // Default value for use mode